create user authorisation for post, partially for comment

+++ magicforum/app/controllers/posts_controller.rb
@@ -3,16 +3,19 @@ class PostsController < ApplicationController
   def index
     @topic = Topic.includes(:posts).find_by(id: params[:topic_id])
     @posts = @topic.posts
+
   end
 
   def new
     @topic = Topic.find_by(id: params[:topic_id])
     @post = Post.new
+    authorize @post
+
   end
 
   def create
     @topic = Topic.find_by(id: params[:topic_id])
-    @post = Post.new(post_params.merge(topic_id: params[:topic_id]))
+    @post = current_user.posts.build(post_params.merge(topic_id: params[:topic_id]))
 
     if @post.save
       flash[:success] = "You've created a new post"
@@ -26,11 +29,13 @@ class PostsController < ApplicationController
   def edit
     @post = Post.find_by(id: params[:id])
     @topic = @post.topic
+    authorize @post
   end
 
   def update
     @post = Post.find_by(id: params[:id])
     @topic = @post.topic
+    authorize @post
 
     if @post.update(post_params)
       flash[:success] = "You've updated post"
@@ -44,6 +49,7 @@ class PostsController < ApplicationController
   def destroy
     @post = Post.find_by(id: params[:id])
     @topic = @post.topic
+    authorize @post
 
     if @post.destroy
       redirect_to topic_posts_path(@topic)
